# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Triangle Curling Spares API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/auth/request-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "contact": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "contact"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "multipleMembers": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success",
                    "multipleMembers"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "contact": {
                    "type": "string",
                    "minLength": 1
                  },
                  "code": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 6
                  }
                },
                "required": [
                  "contact",
                  "code"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "requiresSelection": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "tempToken": {
                          "type": "string"
                        },
                        "members": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": false,
                            "properties": {
                              "id": {
                                "type": "number"
                              },
                              "name": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name"
                            ]
                          }
                        }
                      },
                      "required": [
                        "requiresSelection",
                        "tempToken",
                        "members"
                      ]
                    },
                    {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "token": {
                          "type": "string"
                        },
                        "member": {
                          "type": "object",
                          "additionalProperties": false,
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "name": {
                              "type": "string"
                            },
                            "email": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "phone": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "spareOnly": {
                              "type": "boolean"
                            },
                            "isAdmin": {
                              "type": "boolean"
                            },
                            "isServerAdmin": {
                              "type": "boolean"
                            },
                            "leagueManagerLeagueIds": {
                              "type": "array",
                              "items": {
                                "type": "number"
                              }
                            },
                            "isLeagueAdministrator": {
                              "type": "boolean"
                            },
                            "isLeagueAdministratorGlobal": {
                              "type": "boolean"
                            },
                            "firstLoginCompleted": {
                              "type": "boolean"
                            },
                            "optedInSms": {
                              "type": "boolean"
                            },
                            "emailSubscribed": {
                              "type": "boolean"
                            },
                            "emailVisible": {
                              "type": "boolean"
                            },
                            "phoneVisible": {
                              "type": "boolean"
                            },
                            "themePreference": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "email",
                            "phone",
                            "spareOnly",
                            "isAdmin",
                            "isServerAdmin",
                            "leagueManagerLeagueIds",
                            "isLeagueAdministrator",
                            "isLeagueAdministratorGlobal",
                            "firstLoginCompleted",
                            "optedInSms",
                            "emailSubscribed",
                            "emailVisible",
                            "phoneVisible",
                            "themePreference"
                          ]
                        }
                      },
                      "required": [
                        "token",
                        "member"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/select-member": {
      "post": {
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "memberId": {
                    "type": "number"
                  },
                  "tempToken": {
                    "type": "string"
                  }
                },
                "required": [
                  "memberId",
                  "tempToken"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "requiresSelection": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "tempToken": {
                          "type": "string"
                        },
                        "members": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": false,
                            "properties": {
                              "id": {
                                "type": "number"
                              },
                              "name": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name"
                            ]
                          }
                        }
                      },
                      "required": [
                        "requiresSelection",
                        "tempToken",
                        "members"
                      ]
                    },
                    {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "token": {
                          "type": "string"
                        },
                        "member": {
                          "type": "object",
                          "additionalProperties": false,
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "name": {
                              "type": "string"
                            },
                            "email": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "phone": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "spareOnly": {
                              "type": "boolean"
                            },
                            "isAdmin": {
                              "type": "boolean"
                            },
                            "isServerAdmin": {
                              "type": "boolean"
                            },
                            "leagueManagerLeagueIds": {
                              "type": "array",
                              "items": {
                                "type": "number"
                              }
                            },
                            "isLeagueAdministrator": {
                              "type": "boolean"
                            },
                            "isLeagueAdministratorGlobal": {
                              "type": "boolean"
                            },
                            "firstLoginCompleted": {
                              "type": "boolean"
                            },
                            "optedInSms": {
                              "type": "boolean"
                            },
                            "emailSubscribed": {
                              "type": "boolean"
                            },
                            "emailVisible": {
                              "type": "boolean"
                            },
                            "phoneVisible": {
                              "type": "boolean"
                            },
                            "themePreference": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "email",
                            "phone",
                            "spareOnly",
                            "isAdmin",
                            "isServerAdmin",
                            "leagueManagerLeagueIds",
                            "isLeagueAdministrator",
                            "isLeagueAdministratorGlobal",
                            "firstLoginCompleted",
                            "optedInSms",
                            "emailSubscribed",
                            "emailVisible",
                            "phoneVisible",
                            "themePreference"
                          ]
                        }
                      },
                      "required": [
                        "token",
                        "member"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}